requirements.txt
accelerate==1.2.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
asttokens==3.0.0
attrs==23.2.0
blinker==1.9.0
boto3==1.35.92
botocore==1.35.92
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.0
diffusers==0.32.1
dill==0.3.9
docker==7.1.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.6
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.12.23
fonttools==4.56.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
gunicorn==23.0.0
h11==0.14.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
importlib-metadata==6.11.0
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.3
jmespath==1.0.1
jsonify==0.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==4.0.3
mpmath==1.3.0
multiprocess==0.70.17
narwhals==1.19.1
networkx==3.2.1
numpy==1.26.4
omegaconf==2.2.3
onnx==1.17.0
onnxruntime==1.20.1
onnxruntime-gpu==1.20.1
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathos==0.3.3
pillow==11.0.0
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
prompt_toolkit==3.0.50
protobuf==5.29.2
psutil==6.1.1
pure_eval==0.2.3
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.10.4
safetensors==0.4.5
sagemaker==2.237.1
sagemaker-core==1.0.17
schema==0.7.7
six==1.17.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
streamlit-drawable-canvas==0.9.3
streamlit-image-coordinates==0.2.0
sympy==1.13.1
tblib==3.0.0
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.21.0
